pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "wam2layers"
description = "Atmospheric moisture tracking model"
version = "3.0.0-beta.3"
readme = "README.md"
license = { text = "Apache 2.0" }
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "Topic :: Scientific/Engineering :: Hydrology",
]
keywords = [
  "atmospheric rivers",
  "hydrological cycle",
  "meteorology",
  "moisture recycling",
  "moisture tracking",
]
requires-python = ">=3.8,<3.10"
dependencies = [
    "numpy",
    "pyyaml",
    "scipy",
    "pandas>=1.4.0",
    "xarray", # or xarray[io,parallel,complete,...]
    "netcdf4",
    "scipy",
    "click",
    "pydantic",
    "dask>=2022.11",
    "distributed>=2022.11",
    "psutil",
    "matplotlib",
    "cmocean",
]

[project.optional-dependencies]
develop = [
    "black",
    "cffconvert",
    "isort",
    "myst-nb",
    "sphinx_rtd_theme",
    "sphinx",
    "pytest",
]

[project.urls]
Documentation = "https://wam2layers.readthedocs.io/en/latest"
"Source code" = "https://github.com/WAM2layers/WAM2layers"

[project.scripts]
wam2layers = "wam2layers.cli:cli"

[tool.setuptools.package-data]
"wam2layers.preprocessing" = ["*.csv"]
